requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.13.2
jupyter==1.0.0
notebook==7.0.6
python-dotenv==1.0.0
pydantic==2.8.2

